[DEV] fix missing metadata samples timestamp table
No entries were written for the metadata track STTS
(Decoding Time to Sample Box).
The metadata timestamps are the same than the video frames timestamp
but some frames may not have metadata, therefore the video track STTS
cannot be simply copied for the metadata track and a specific table
must be written.

Change-Id: I9ce30a332148163e0df2c5b4c5ecbd97c3c63794
Reviewed-on: https://gerrit.parrot.biz/49375
Reviewed-by: Jean-Baptiste Dubois <jean-baptiste.dubois@parrot.com>