[DEV] make video frames buffer read-only
The library must not write to the frames buffer as the frame may not be
used exclusively by the library.
Previously the H.264 NAL unit start codes were replaced by NAL units size
in the user frame buffer prior to writing the frame data to the file.
Now H.264 NAL unit sizes are written first to the file and then the NALU
data (with a 4 bytes offset).

Change-Id: I2f816014e43f85fe8a30569031aa16377369c77a
Reviewed-on: https://gerrit.parrot.biz/49409
Reviewed-by: Jean-Baptiste Dubois <jean-baptiste.dubois@parrot.com>